+++ docs/make.jl
@@ -12,7 +12,6 @@ makedocs(
         "Overview" => "index.md",
         "ParticleSystem interface" => "ParticleSystem.md",
         "Neighbor lists" => "neighborlists.md",
-        "Low level interface" => "LowLevel.md",
         "Unitcell requirements" => "unitcell.md",
         "Ecosystem integration" => "ecosystem.md",
         "From Python" => "python.md",